delegate functionality from FakeRoot to FakeBucket
We need the header-based access method for usage with s3fs. Maybe
we can model FakeRoot as subclass of FakeBucket in the future (this
should remove all the 'forward-only' methods). I'd also like to check
against the presence of the bucket header within the HTTP request (
currently not done, as we only have one fake bucket)